+++ src/snf_simulations/data/fispin.py
